+++ source/ProcessManager.Client.Tests/ProcessManager.Client.Tests.csproj
@@ -8,13 +8,15 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Include="..\Example.Orchestrations.Tests\Fixtures\ExampleOrchestrationsAppManager.cs" Link="Fixtures\ExampleOrchestrationsAppManager.cs" />
     <Compile Include="..\ProcessManager.Core.Tests\Fixtures\ProcessManagerDatabaseManager.cs" Link="Fixtures\ProcessManagerDatabaseManager.cs" />
     <Compile Include="..\ProcessManager.Orchestrations.Tests\Fixtures\OrchestrationsAppManager.cs">
       <Link>Fixtures\OrchestrationsAppManager.cs</Link>
     </Compile>
     <Compile Include="..\ProcessManager.Tests\Fixtures\ProcessManagerAppManager.cs">
       <Link>Fixtures\ProcessManagerAppManager.cs</Link>
+    </Compile>
+    <Compile Include="..\ProcessManagerExample.Orchestrations.Tests\Fixtures\ExampleOrchestrationsAppManager.cs">
+      <Link>Fixtures\ExampleOrchestrationsAppManager.cs</Link>
     </Compile>
   </ItemGroup>
 
@@ -33,7 +35,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Example.Orchestrations\Example.Orchestrations.csproj" />
+    <ProjectReference Include="..\ProcessManager.Example.Orchestrations\ProcessManager.Example.Orchestrations.csproj" />
     <ProjectReference Include="..\ProcessManager.Client\ProcessManager.Client.csproj" />
     <ProjectReference Include="..\ProcessManager.DatabaseMigration\ProcessManager.DatabaseMigration.csproj" />
     <ProjectReference Include="..\ProcessManager.Orchestrations\ProcessManager.Orchestrations.csproj" />
