+++ source/ProcessManager.Example.Orchestrations.Abstrations/ProcessManager.Example.Orchestrations.Abstrations.csproj
