+++ source/ProcessManager.Example.Orchestrations.Abstrations/Processes/BRS_X01/InputExample/V1/Model/Brs_X01_InputExample_V1.cs
