+++ source/ProcessManager.Example.Orchestrations.Abstrations/Processes/BRS_X01/InputExample/V1/Model/StartInputExampleCommandV1.cs
