+++ source/ProcessManager.Example.Orchestrations.Abstrations/Processes/BRS_X01/NoInputExample/V1/Model/Brs_X01_NoInputExample_V1.cs
