+++ source/ProcessManager.Example.Orchestrations.Abstrations/Processes/BRS_X01/NoInputExample/V1/Model/StartNoInputExampleCommandV1.cs
