+++ source/ProcessManager.Example.Orchestrations/Processes/BRS_X01/InputExample/SearchTrigger_Brs_X01_InputExample.cs
