+++ source/ProcessManager.Example.Orchestrations/Processes/BRS_X01/InputExample/V1/Activities/InitializeOrchestrationActivity_Brs_X01_InputExample_V1.cs
