+++ source/ProcessManager.Example.Orchestrations/Processes/BRS_X01/InputExample/V1/Activities/ProgressActivityBase.cs
