+++ source/ProcessManager.Example.Orchestrations/Processes/BRS_X01/NoInputExample/V1/Activities/FirstStepStopActivity_Brs_X01_NoInputExample_V1.cs
