+++ source/ProcessManager.Example.Orchestrations/Processes/BRS_X01/NoInputExample/V1/Activities/ProgressActivityBase.cs
