+++ source/ProcessManagerExample.Orchestrations.Tests/Fixtures/ExampleOrchestrationsAppCollection.cs
